Fixed suspending issue

+++ Apps.GoogleDrive/Webhooks/Handlers/ChangesHandler.cs
@@ -3,13 +3,14 @@ using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Authentication;
 using Blackbird.Applications.Sdk.Common.Invocation;
 using Blackbird.Applications.Sdk.Common.Webhooks;
-using Blackbird.Applications.Sdk.Utils.Webhooks.Bridge;
 using Google.Apis.Drive.v3.Data;
 
 namespace Apps.GoogleDrive.Webhooks.Handlers;
 
 public class ChangesHandler : BaseInvocable, IWebhookEventHandler
 {
+    private const string StoredValueNotFound = "Stored value was not found";
+    
     public ChangesHandler(InvocationContext invocationContext) : base(invocationContext)
     {
     }
@@ -34,7 +35,7 @@ public class ChangesHandler : BaseInvocable, IWebhookEventHandler
         var stateToken = client.Changes.GetStartPageToken().Execute();
         var bridgeService = new BridgeService(InvocationContext.UriInfo.BridgeServiceUrl.ToString());
         await bridgeService.StoreValue(InvocationContext.Bird.Id.ToString(), stateToken.StartPageTokenValue);
-
+        
         var request = client.Changes.Watch(channel, stateToken.StartPageTokenValue);
         await request.ExecuteAsync();
     }
@@ -43,6 +44,12 @@ public class ChangesHandler : BaseInvocable, IWebhookEventHandler
     {
         var bridgeService = new BridgeService(InvocationContext.UriInfo.BridgeServiceUrl.ToString());
         var resourceId = (await bridgeService.RetrieveValue(InvocationContext.Bird.Id.ToString() + "_resourceId")).Replace("\"", "");
+        if (resourceId == StoredValueNotFound)
+        {
+            // If resource id is not found, there is no need to unsubscribe
+            return;
+        }
+        
         await bridgeService.DeleteValue(InvocationContext.Bird.Id.ToString() + "_resourceId");
 
         var client = new GoogleDriveClient(authenticationCredentialsProvider);
@@ -51,6 +58,7 @@ public class ChangesHandler : BaseInvocable, IWebhookEventHandler
             Id = InvocationContext.Bird.Id.ToString(),
             ResourceId = resourceId
         };
+        
         var request = client.Channels.Stop(channel);
         await request.ExecuteAsync();
     }
